refactor: ContainerDeleter created
