+++ src/library-authoring/components/ContainerCard.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useContext, ReactNode } from 'react';
+import { useCallback, ReactNode } from 'react';
 import { FormattedMessage, useIntl } from '@edx/frontend-platform/i18n';
 import {
   ActionRow,
@@ -8,7 +8,7 @@ import {
   useToggle,
   Stack,
 } from '@openedx/paragon';
-import { MoreVert, Warning } from '@openedx/paragon/icons';
+import { MoreVert } from '@openedx/paragon/icons';
 import { Link } from 'react-router-dom';
 
 import { getItemIcon, getComponentStyleColor } from '../../generic/block-type-utils';
@@ -19,9 +19,8 @@ import { useSidebarContext } from '../common/context/SidebarContext';
 import BaseCard from './BaseCard';
 import { useLibraryRoutes } from '../routes';
 import messages from './messages';
-import DeleteModal from '../../generic/delete-modal/DeleteModal';
-import { ToastContext } from '../../generic/toast-context';
-import { useDeleteContainer, useRestoreContainer, useContainerChildren } from '../data/apiHooks';
+import { useContainerChildren } from '../data/apiHooks';
+import ContainerDeleter from './ContainerDeleter';
 
 type ContainerMenuProps = {
   hit: ContainerHit,
@@ -32,88 +31,44 @@ const ContainerMenu = ({ hit } : ContainerMenuProps) => {
   const {
     contextKey,
     blockId,
-    usageKey: containerKey,
+    usageKey: containerId,
     displayName,
   } = hit;
-  const {
-    sidebarComponentInfo,
-    closeLibrarySidebar,
-  } = useSidebarContext();
 
   const [isConfirmingDelete, confirmDelete, cancelDelete] = useToggle(false);
-  const { showToast } = useContext(ToastContext);
-  const deleteContainerMutation = useDeleteContainer(containerKey);
-
-  const deleteWarningTitle = intl.formatMessage(messages.deleteUnitWarningTitle);
-  const deleteText = intl.formatMessage(messages.deleteUnitConfirm, {
-    componentName: displayName,
-  });
-  const deleteSuccess = intl.formatMessage(messages.deleteUnitSuccess);
-  const deleteError = intl.formatMessage(messages.deleteUnitFailed);
-  const undoDeleteError = messages.undoDeleteUnitToastFailed;
-
-  const restoreContainerMutation = useRestoreContainer(containerKey);
-  const restoreComponent = useCallback(async () => {
-    try {
-      await restoreContainerMutation.mutateAsync();
-      showToast(intl.formatMessage(messages.undoDeleteContainerToastMessage));
-    } catch (e) {
-      showToast(intl.formatMessage(undoDeleteError));
-    }
-  }, []);
-
-  const onDelete = useCallback(async () => {
-    await deleteContainerMutation.mutateAsync().then(() => {
-      if (sidebarComponentInfo?.id === containerKey) {
-        closeLibrarySidebar();
-      }
-      showToast(
-        deleteSuccess,
-        {
-          label: intl.formatMessage(messages.undoDeleteContainerToastAction),
-          onClick: restoreComponent,
-        },
-      );
-    }).catch(() => {
-      showToast(deleteError);
-    }).finally(() => {
-      cancelDelete();
-    });
-  }, [sidebarComponentInfo, showToast, deleteContainerMutation]);
 
   return (
-    <Dropdown id="container-card-dropdown">
-      <Dropdown.Toggle
-        id="container-card-menu-toggle"
-        as={IconButton}
-        src={MoreVert}
-        iconAs={Icon}
-        variant="primary"
-        alt={intl.formatMessage(messages.collectionCardMenuAlt)}
-        data-testid="container-card-menu-toggle"
-      />
-      <Dropdown.Menu>
-        <Dropdown.Item
-          as={Link}
-          to={`/library/${contextKey}/container/${blockId}`}
-          disabled
-        >
-          <FormattedMessage {...messages.menuOpen} />
-        </Dropdown.Item>
-        <Dropdown.Item onClick={confirmDelete}>
-          <FormattedMessage {...messages.menuDeleteContainer} />
-        </Dropdown.Item>
-      </Dropdown.Menu>
-      <DeleteModal
+    <>
+      <Dropdown id="container-card-dropdown">
+        <Dropdown.Toggle
+          id="container-card-menu-toggle"
+          as={IconButton}
+          src={MoreVert}
+          iconAs={Icon}
+          variant="primary"
+          alt={intl.formatMessage(messages.collectionCardMenuAlt)}
+          data-testid="container-card-menu-toggle"
+        />
+        <Dropdown.Menu>
+          <Dropdown.Item
+            as={Link}
+            to={`/library/${contextKey}/container/${blockId}`}
+            disabled
+          >
+            <FormattedMessage {...messages.menuOpen} />
+          </Dropdown.Item>
+          <Dropdown.Item onClick={confirmDelete}>
+            <FormattedMessage {...messages.menuDeleteContainer} />
+          </Dropdown.Item>
+        </Dropdown.Menu>
+      </Dropdown>
+      <ContainerDeleter
         isOpen={isConfirmingDelete}
         close={cancelDelete}
-        variant="warning"
-        title={deleteWarningTitle}
-        icon={Warning}
-        description={deleteText}
-        onDeleteSubmit={onDelete}
+        containerId={containerId}
+        displayName={displayName}
       />
-    </Dropdown>
+    </>
   );
 };
